+++ src/main/java/com/guillermomolina/protos/parser/ast/SurfaceExpression.java
@@ -34,6 +34,7 @@ public sealed interface SurfaceExpression
                 SurfaceSlotCreation,
                 SurfaceAssignment,
                 SurfaceSuperSend,
-                SurfaceObject {
+                SurfaceObject,
+                SurfaceClosure {
     SourceSpan span();
 }